requirements.txt
torch>=2.0.0
torchvision
torchaudio
transformers>=4.38.0
datasets
accelerate
pillow
huggingface_hub
sentencepiece
jupyter
ipykernel
matplotlib
timm
streamlit
huggingface_hub
typer
evaluate
rouge-score
nltk
pycocoevalcapqwen-vl-utils
qwen-vl-utils
